Compares and combines 2-group proportion ratios in original and follow-up studies

This function computes confidence intervals from an original study and a follow-up study where the effect size is a 2-group proportion ratio. Confidence intervals for the ratio and geometric average of effect sizes are also computed. The confidence level for the ratio of ratios is 1 – 2*alpha, which is recommended for equivalence testing.

replicate.ratio.prop2(alpha, f11, f12, n11, n12, f21, f22, n21, n22)

Arguments

alpha

alpha level for 1-alpha confidence 21q``````````````````````` `````````````````````````````````````````````````````````````````````````````````````````````````````````````````````````````````````````````````````````````````````````````````````````````````````lpha alpha level for 1-alpha confidence

f11

frequency count for group 1 in original study

f12

frequency count for group 2 in original study

n11

sample size for group 1 in original study

n12

sample size for group 2 in original study

f21

frequency count for group 1 in follow-up study

f22

frequency count for group 2 in follow-up study

n21

sample size for group 1 in follow-up study

n22

sample size for group 2 in follow-up study

Value

A 4-row matrix. The rows are:

Row 1 summarizes the original study

Row 2 summarizes the follow-up study

Row 3 estimates the ratio of proportion ratios

Row 4 estimates the geometric average proportion ratio

The columns are:

Estimate - proportion difference estimate (single study, ratio, average)

LL - lower limit of the confidence interval

UL - upper limit of the confidence interval

References

Bonett DG (2021). “Design and analysis of replication studies.” Organizational Research Methods, 24(3), 513--529. ISSN 1094-4281, doi:10.1177/1094428120911088 .

Examples

replicate.ratio.prop2(.05, 21, 16, 40, 40, 19, 13, 60, 60)
#>                      Estimate        LL       UL
#> Original:           1.3076923 0.8068705 2.119373
#> Follow-up:          1.4528302 0.7939881 2.658372
#> Original/Follow-up: 0.9000999 0.4703209 1.722611
#> Average:            1.3783522 0.9362893 2.029132

# Should return:
#                      Estimate        LL       UL
# Original:           1.3076923 0.8068705 2.119373
# Follow-up:          1.4528302 0.7939881 2.658372
# Original/Follow-up: 0.9000999 0.4703209 1.722611
# Average:            1.3783522 0.9362893 2.029132