Computes the standard error for a paired-samples mean difference

This function computes the standard error of a paired-samples mean difference using the estimated means, estimated standard deviations, estimated Pearson correlation, and sample size. The effect size estimate and standard error output from this function can be used as input in the meta.ave.gen, meta.lc.gen, and meta.lm.gen functions in applications where compatible mean differences from a combination of 2-group and paired-samples experiments are used in the meta-analysis. Equality of variances is not assumed.

se.mean.ps(m1, m2, sd1, sd2, cor, n)

Arguments

m1

estimated mean for measurement 1

m2

estimated mean for measurement 2

sd1

estimated standard deviation for measurement 1

sd2

estimated standard deviation for measurement 2

cor

estimated correlation for measurements 1 and 2

n

sample size

Value

Returns a one-row matrix:

Estimate - estimated mean difference

SE - standard error

References

Snedecor GW, Cochran WG (1980). Statistical methods, 7th edition. ISU University Pres, Ames, Iowa.

Examples

se.mean.ps(23.9, 25.1, 1.76, 2.01, .78, 25)
#>                   Estimate        SE
#> Mean difference:      -1.2 0.2544833

# Should return:
#                   Estimate        SE
# Mean difference:      -1.2 0.2544833