Computes the estimate and standard error for a paired-samples log proportion ratio

Computes a large-sample standard error of a paired-samples log proportion ratio using the frequency counts from a 2 x 2 contingency table. The log proportion ratio and standard error output from this function can be used as input in the meta.ave.gen.log function in applications where compatible proportion ratios from a combination of 2-group and paired-samples studies are used in the meta-analysis.

Usage

se.propratio.ps(f00, f01, f10, f11)

Arguments

f00

number of participants with y = 0 and x = 0

f01

number of participants with y = 0 and x = 1

f10

number of participants with y = 1 and x = 0

f11

number of participants with y = 1 and x = 1

Value

Returns a one-row matrix:

Estimate - estimated log proportion ratio

SE - standard error

Examples

se.propratio.ps(16, 64, 5, 15)
#>                         Estimate        SE
#> Log proportion ratio:  -1.373716 0.2089758

# Should return:
#                         Estimate         SE
# Log proportion ratio:  -1.373716  0.2089758