Confidence interval for an average correlation of any type

Computes the estimate, standard error, and confidence interval for an average correlation. Any type of correlation can be used (e.g., Pearson, Spearman, semipartial, factor correlation, gamma coefficient, Somers d coefficient, tetrachoric, point-biserial, biserial, correlation between latent factors, etc.).

Usage

meta.ave.cor.gen(alpha, cor, se, bystudy = TRUE)

Arguments

alpha

alpha level for 1-alpha confidence

cor

vector of estimated correlations

se

vector of standard errors

bystudy

logical to also return each study estimate (TRUE) or not

Value

Returns a matrix. The first row is the average estimate across all studies. If bystudy is TRUE, there is 1 additional row for each study. The matrix has the following columns:

Estimate - estimated effect size

SE - standard error

LL - lower limit of the confidence interval

UL - upper limit of the confidence interval

References

Bonett DG (2008). “Meta-analytic interval estimation for bivariate correlations.” Psychological Methods, 13(3), 173–181. ISSN 1939-1463, doi:10.1037/a0012868 .

Examples

cor <- c(.396, .454, .409, .502, .350)
se <- c(.104, .064, .058, .107, .086)
meta.ave.cor.gen(.05, cor, se, bystudy = TRUE)
#>         Estimate         SE        LL        UL
#> Average   0.4222 0.03853362 0.3438560 0.4947070
#> Study 1   0.3960 0.10400000 0.1753200 0.5787904
#> Study 2   0.4540 0.06400000 0.3200675 0.5701415
#> Study 3   0.4090 0.05800000 0.2893856 0.5160375
#> Study 4   0.5020 0.10700000 0.2651183 0.6817343
#> Study 5   0.3500 0.08600000 0.1716402 0.5061435

# Should return:
#         Estimate         SE        LL        UL
# Average   0.4222 0.03853362 0.3438560 0.4947070
# Study 1   0.3960 0.10400000 0.1753200 0.5787904
# Study 2   0.4540 0.06400000 0.3200675 0.5701415
# Study 3   0.4090 0.05800000 0.2893856 0.5160375
# Study 4   0.5020 0.10700000 0.2651183 0.6817343
# Study 5   0.3500 0.08600000 0.1716402 0.5061435