Confidence interval for an average Cronbach alpha reliability

Computes the estimate, standard error, and confidence interval for an average Cronbach reliability coefficient from two or more studies.

meta.ave.cronbach(alpha, n, rel, r, bystudy = TRUE)

Arguments

alpha

alpha level for 1-alpha confidence

n

vector of sample sizes

rel

vector of sample reliabilities

r

number of measurements (e.g., items) used to compute each reliability

bystudy

logical to also return each study estimate (TRUE) or not

Value

Returns a matrix. The first row is the average estimate across all studies. If bystudy is true, there is 1 additional row for each study. The matrix has the following columns:

Estimate - the estimated effect size

SE - standard error

LL - lower limit of the confidence interval

UL - upper limit of the confidence interval

References

Bonett DG (2010). “Varying coefficient meta-analytic methods for alpha reliability.” Psychological Methods, 15(4), 368--385. ISSN 1939-1463, doi: 10.1037/a0020142 , https://doi.apa.org/getdoi.cfm?doi=10.1037/a0020142.

Examples

n <- c(583, 470, 546, 680)
rel <- c(.91, .89, .90, .89)
meta.ave.cronbach(.05, n, rel, 10, bystudy = TRUE)
#>         Estimate          SE        LL        UL
#> Average   0.8975 0.003256081 0.8911102 0.9038592
#> Study 1   0.9100 0.005566064 0.8985763 0.9204108
#> Study 2   0.8900 0.007579900 0.8743616 0.9041013
#> Study 3   0.9000 0.006391375 0.8868623 0.9119356
#> Study 4   0.8900 0.006297549 0.8771189 0.9018203

# Should return:
#         Estimate          SE        LL        UL
# Average   0.8975 0.003256081 0.8911102 0.9038592
# Study 1   0.9100 0.005566064 0.8985763 0.9204108
# Study 2   0.8900 0.007579900 0.8743616 0.9041013
# Study 3   0.9000 0.006391375 0.8868623 0.9119356
# Study 4   0.8900 0.006297549 0.8771189 0.9018203