Confidence interval for an average point-biserial correlation

Computes the estimate, standard error, and confidence interval for an average point-biserial correlation from two or more studies. Two types of point-biserial correlations can be meta-analyzed. One type uses an unweighted variance and is appropriate in 2-group experimental designs. The other type uses a weighted variance and is appropriate in 2-group nonexperimental designs with simple random sampling within each group. This function requires all point-biserial correlations to be of the same type. Use the meta.ave.gen function to meta-analyze any combination of biserial correlation types.

meta.ave.pbcor(alpha, m1, m2, sd1, sd2, n1, n2, type, bystudy = TRUE)

Arguments

alpha
alpha level for 1-alpha confidence

m1
vector of sample means for group 1

m2
vector of sample means for group 2

sd1
vector of sample SDs for group 1

sd2
vector of sample SDs for group 2

n1
vector of group 1 sample sizes

n2
vector of group 2 sample sizes

type

set to 1 for weighted variance

set to 2 for unweighted variance

bystudy
logical to also return each study estimate (TRUE) or not

Value

Returns a matrix. The first row is the average estimate across all studies. If bystudy is true, there is 1 additional row for each study. The matrix has the following columns:

Estimate - the estimated effect size

SE - standard error

LL - lower limit of the confidence interval

UL - upper limit of the confidence interval

References

Bonett DG (2020). “Point-biserial correlation: Interval estimation, hypothesis testing, meta-analysis, and sample size determination.” British Journal of Mathematical and Statistical Psychology, 73(S1), 113--144. ISSN 0007-1102, doi: 10.1111/bmsp.12189 , https://onlinelibrary.wiley.com/doi/10.1111/bmsp.12189.

Examples

m1 <- c(21.9, 23.1, 19.8)
m2 <- c(16.1, 17.4, 15.0)
sd1 <- c(3.82, 3.95, 3.67)
sd2 <- c(3.21, 3.30, 3.02)
n1 <- c(40, 30, 24)
n2 <- c(40, 28, 25)
meta.ave.pbcor(.05, m1, m2, sd1, sd2, n1, n2, 2, bystudy = TRUE)

#>          Estimate         SE        LL        UL
#> Average 0.6108874 0.04531854 0.5143535 0.6921339
#> Study 1 0.6349786 0.06316796 0.4842098 0.7370220
#> Study 2 0.6165315 0.07774063 0.4260068 0.7384924
#> Study 3 0.5811521 0.09192459 0.3550766 0.7236206

# Should return:
#          Estimate         SE        LL        UL
# Average 0.6159094 0.04363432 0.5230976 0.6942842
# Study 1 0.6349786 0.06316796 0.4842098 0.7370220
# Study 2 0.6160553 0.07776700 0.4255342 0.7380898
# Study 3 0.5966942 0.08424778 0.3903883 0.7283966