Confidence interval for an average Spearman correlation

Computes the estimate, standard error, and confidence interval for an average Spearman correlation from two or more studies. The Spearman correlation is preferred to the Pearson correlation if the relation between the two quantitative variables is monotonic rather than linear or if the bivariate normality assumption is not plausible.

Usage

meta.ave.spear(alpha, n, cor, bystudy = TRUE)

Arguments

alpha

alpha level for 1-alpha confidence

n

vector of sample sizes

cor

vector of estimated Spearman correlations

bystudy

logical to also return each study estimate (TRUE) or not

Value

Returns a matrix. The first row is the average estimate across all studies. If bystudy is TRUE, there is 1 additional row for each study. The matrix has the following columns:

Estimate - estimated effect size

SE - standard error

LL - lower limit of the confidence interval

UL - upper limit of the confidence interval

References

Bonett DG (2008). “Meta-analytic interval estimation for bivariate correlations.” Psychological Methods, 13(3), 173–181. ISSN 1939-1463, doi:10.1037/a0012868 .

Examples

n <- c(150, 200, 300, 200, 350)
cor <- c(.14, .29, .16, .21, .23)
meta.ave.spear(.05, n, cor, bystudy = TRUE)
#>         Estimate         SE          LL        UL
#> Average    0.206 0.02944265  0.14763960 0.2629309
#> Study 1    0.140 0.08071009 -0.02151639 0.2943944
#> Study 2    0.290 0.06627745  0.15476515 0.4145671
#> Study 3    0.160 0.05671051  0.04689807 0.2690514
#> Study 4    0.210 0.06850496  0.07187439 0.3402225
#> Study 5    0.230 0.05136319  0.12690280 0.3281809

# Should return:
#         Estimate         SE          LL        UL
# Average    0.206 0.02944265  0.14763960 0.2629309
# Study 1    0.140 0.08031750 -0.02151639 0.2943944
# Study 2    0.290 0.06492643  0.15476515 0.4145671
# Study 3    0.160 0.05635101  0.04689807 0.2690514
# Study 4    0.210 0.06776195  0.07187439 0.3402225
# Study 5    0.230 0.05069710  0.12690280 0.3281809