Confidence interval for an average standardized mean difference from 2-group studies

Computes the estimate, standard error, and confidence interval for an average standardized mean difference from two or more 2-group studies. Unweighted variances, weighted variances, and single group variance are options for the standardizer. Equal variances within or across studies is not assumed.

meta.ave.stdmean2(alpha, m1, m2, sd1, sd2, n1, n2, stdzr, bystudy = TRUE)

Arguments

alpha
alpha level for 1-alpha confidence

m1
vector of sample means for group 1

m2
vector of sample means for group 2

sd1
vector of sample SDs for group 1

sd2
vector of sample SDs for group 2

n1
vector of group 1 sample sizes

n2
vector of group 2 sample sizes

stdzr

set to 0 for square root average variance standardizer

set to 1 for group 1 SD standardizer

set to 2 for group 2 SD standardizer

set to 3 for square root weighted variance standardizer

bystudy
logical to also return each study estimate (TRUE) or not

Value

A matrix. First row is the overall average estimate. If bystudy is TRUE also returns 1 row per study. The matrix has the following columns:

Estimate - the estimated effect size

SE - standard error

LL - lower limit of the confidence interval

UL - upper limit of the confidence interval

References

Bonett DG (2009). “Meta-analytic interval estimation for standardized and unstandardized mean differences.” Psychological Methods, 14(3), 225--238. ISSN 1939-1463, doi: 10.1037/a0016619 , https://doi.apa.org/getdoi.cfm?doi=10.1037/a0016619.

Examples

m1 <- c(21.9, 23.1, 19.8)
m2 <- c(16.1, 17.4, 15.0)
sd1 <- c(3.82, 3.95, 3.67)
sd2 <- c(3.21, 3.30, 3.02)
n1 <- c(40, 30, 24)
n2 <- c(40, 28, 25)
meta.ave.stdmean2(.05, m1, m2, sd1, sd2, n1, n2, 0, bystudy = TRUE)

#>         Estimate        SE        LL       UL
#> Average 1.526146 0.1734341 1.1862217 1.866071
#> Study 1 1.643894 0.2629049 1.1286100 2.159178
#> Study 2 1.566132 0.3056278 0.9671126 2.165152
#> Study 3 1.428252 0.3289179 0.7835848 2.072919

# Should return: 
#         Estimate        SE        LL       UL
# Average 1.526146 0.1734341 1.1862217 1.866071
# Study 1 1.643894 0.2629049 1.1286100 2.159178
# Study 2 1.566132 0.3056278 0.9671126 2.165152
# Study 3 1.428252 0.3289179 0.7835848 2.072919