Compares Pearson or partial correlations in original and follow-up studies

This function can be used to compare and combine Pearson or partial correlations from the original study and the follow-up study. The confidence level for the difference is 1 – 2alpha.

replicate.cor(alpha, cor1, n1, cor2, n2, s)

Arguments

alpha
alpha level for 1-alpha confidence

cor1
sample Pearson correlation between y and x in original study

n1
sample size in original study

cor2
sample Pearson correlation between y and x in follow-up study

n2
sample size in follow-up study

s
number of control variables in each study (0 for Pearson)

Value

A 4-row matrix. The rows are:

Row 1 summarizes the Original study

Row 2 summarizes the Follow-up study

Row 3 estimates the difference between studies

Row 4 estimates the average effect size between the two studies

The columns are:

Estimate - Pearson or partial correlation estimate

SE - standard error

z - z-value

p - p-value

LL - lower limit of the confidence interval

UL - upper limit of the confidence interval

References

Bonett DG (2021). “Design and analysis of replication studies.” Organizational Research Methods, 24(3), 513--529. ISSN 1094-4281, doi: 10.1177/1094428120911088 , https://journals.sagepub.com/doi/10.1177/1094428120911088.

Examples

replicate.cor(.05, .598, 80, .324, 200, 0)

#>                       Estimate         SE        z            p        LL
#> Original:                0.598 0.11396058 6.589418 4.708045e-09 0.4355043
#> Follow-up:               0.324 0.07124705 4.819037 2.865955e-06 0.1939787
#> Original - Follow-up:    0.274 0.09708614 2.633335 8.455096e-03 0.1065496
#> Average:                 0.461 0.04854307 7.634998 2.264855e-14 0.3725367
#>                              UL
#> Original:             0.7227538
#> Follow-up:            0.4428347
#> Original - Follow-up: 0.4265016
#> Average:              0.5411607

# Should return:
#                      Estimate         SE        z            p        LL        UL
#  Original:              0.598 0.11396058 6.589418 4.708045e-09 0.4355043 0.7227538
#  Follow-up:             0.324 0.07124705 4.819037 2.865955e-06 0.1939787 0.4428347
#  Original - Follow-up:  0.274 0.09708614 2.633335 8.455096e-03 0.1065496 0.4265016
#  Average:               0.461 0.04854307 7.634998 2.264855e-14 0.3725367 0.5411607