Compares and combines Cronbach reliablity in original and follow-up studies

This function can be used to compare and combine a Cronbach reliablity coefficient from an original study and a follow-up study. The confidence level for the difference is 1 – 2*alpha, which is recommended for equivalence testing.

Usage

replicate.cronbach(alpha, rel1, n1, rel2, n2, r)

Arguments

alpha

alpha level for 1-alpha confidence

rel1

estimated reliability in original study

n1

sample size in original study

rel2

estimated reliability in follow-up study

n2

sample size in follow-up study

r

number of measurements (e.g., items)

Value

A 4-row matrix. The rows are:

Row 1 summarizes the original study

Row 2 summarizes the follow-up study

Row 3 estimates the difference in correlations

Row 4 estimates the average correlation

The columns are:

Estimate - correlation estimate (single study, difference, average)

SE - standard error

LL - lower limit of the confidence interval

UL - upper limit of the confidence interval

References

Bonett DG (2010). “Varying coefficient meta-analytic methods for alpha reliability.” Psychological Methods, 15(4), 368–385. ISSN 1939-1463, doi:10.1037/a0020142 . Bonett DG, Price RM (2015). “Varying coefficient meta-analysis methods for odds ratios and risk ratios.” Psychological Methods, 20(3), 394–406. ISSN 1939-1463, doi:10.1037/met0000032 . Bonett DG (2021). “Design and analysis of replication studies.” Organizational Research Methods, 24(3), 513–529. ISSN 1094-4281, doi:10.1177/1094428120911088 .

Examples

replicate.cronbach(.05, .883, 100, .869, 200, 6)
#>                       Estimate         SE          LL         UL
#> Original:                0.883 0.01830958  0.84356871 0.91522517
#> Follow-up:               0.869 0.01442263  0.83874629 0.89523760
#> Original - Follow-up:    0.014 0.02330779 -0.03336284 0.05820123
#> Average:                 0.876 0.01172239  0.85187755 0.89774525

# Should return:
#                       Estimate         SE          LL         UL
# Original:                0.883 0.01830958  0.84356871 0.91522517
# Follow-up:               0.869 0.01442263  0.83874629 0.89523760
# Original - Follow-up:    0.014 0.02330779 -0.03336284 0.05820123
# Average:                 0.876 0.01172239  0.85187755 0.89774525