Compares effect sizes in the original and follow-up studies

This function can be used to compare and combine any effect size (e.g.,odds ratio, proportion ratio, proportion difference, slope coefficient, etc.) using the effect size estimate and its standard error from the original study and the follow-up study. The same results can be obtained using the meta.lc.gen function with appropriate contrast coefficients. The confidence level for the difference is 1 – 2alpha.

replicate.gen(alpha, est1, se1, est2, se2)

Arguments

alpha
alpha level for 1-alpha confidence

est1
sample effect size in original study

se1
effect size standard error in original study

est2
sample effect size in follow-up study

se2
effect size standard error in follow-up study

Value

A 4-row matrix. The rows are:

Row 1 summarizes the Original study

Row 2 summarizes the Follow-up study

Row 3 estimates the difference between studies

Row 4 estimates the average effect size between the two studies

Columns are:

Estimate - effect size estimate

SE - standard error

z - z-value

p - p-value

LL - lower limit of the confidence interval

UL - upper limit of the confidence interval

References

Bonett DG (2021). “Design and analysis of replication studies.” Organizational Research Methods, 24(3), 513--529. ISSN 1094-4281, doi: 10.1177/1094428120911088 , https://journals.sagepub.com/doi/10.1177/1094428120911088.

Examples

replicate.gen(.05, .782, .210, .650, .154)

#>                       Estimate        SE         z            p         LL
#> Original:                0.782 0.2100000 3.7238095 1.962390e-04  0.3704076
#> Follow-up:               0.650 0.1540000 4.2207792 2.434593e-05  0.3481655
#> Original - Follow-up:    0.132 0.2604151 0.5068831 6.122368e-01 -0.2963446
#> Average:                 0.716 0.1302075 5.4989141 3.821373e-08  0.4607979
#>                              UL
#> Original:             1.1935924
#> Follow-up:            0.9518345
#> Original - Follow-up: 0.5603446
#> Average:              0.9712021

# Should return: 
#                      Estimate        SE         z            p         LL        UL
#  Original:              0.782 0.2100000 3.7238095 1.962390e-04  0.3704076 1.1935924
#  Follow-up:             0.650 0.1540000 4.2207792 2.434593e-05  0.3481655 0.9518345
#  Original - Follow-up:  0.132 0.2604151 0.5068831 6.122368e-01 -0.2963446 0.5603446
#  Average:               0.716 0.1302075 5.4989141 3.821373e-08  0.4607979 0.9712021