Compares and combines Spearman correlations in original and follow-up studies

This function can be used to compare and combine Spearman correlations from an original study and a follow-up study. The confidence level for the difference is 1 – 2*alpha, which is recommended for equivalence testing.

Usage

replicate.spear(alpha, cor1, n1, cor2, n2)

Arguments

alpha

alpha level for 1-alpha confidence

cor1

estimated Spearman correlation in original study

n1

sample size in original study

cor2

estimated Spearman correlation in follow-up study

n2

sample size in follow-up study

Value

A 4-row matrix. The rows are:

Row 1 summarizes the original study

Row 2 summarizes the follow-up study

Row 3 estimates the difference in correlations

Row 4 estimates the average correlation

The columns are:

Estimate - Spearman correlation estimate (single study, difference, average)

SE - standard error

z - z-value

p - p-value

LL - lower limit of the confidence interval

UL - upper limit of the confidence interval

References

Bonett DG (2021). “Design and analysis of replication studies.” Organizational Research Methods, 24(3), 513–529. ISSN 1094-4281, doi:10.1177/1094428120911088 .

Examples

replicate.spear(.05, .598, 80, .324, 200)
#>                       Estimate         SE        z            p         LL
#> Original:                0.598 0.07948367 5.315140 1.065752e-07 0.41985966
#> Follow-up:               0.324 0.06541994 4.570582 4.863705e-06 0.19049455
#> Original - Follow-up:    0.274 0.10294378 3.437975 5.860809e-04 0.09481418
#> Average:                 0.461 0.05147189 9.967944 0.000000e+00 0.36695230
#>                              UL
#> Original:             0.7317733
#> Follow-up:            0.4457384
#> Original - Follow-up: 0.4342171
#> Average:              0.5457190

# Should return:
#                       Estimate         SE        z            p         LL        UL
# Original:                0.598 0.07948367 5.315140 1.065752e-07 0.41985966 0.7317733
# Follow-up:               0.324 0.06541994 4.570582 4.863705e-06 0.19049455 0.4457384
# Original - Follow-up:    0.274 0.10294378 3.437975 5.860809e-04 0.09481418 0.4342171
# Average:                 0.461 0.05147189 9.967944 0.000000e+00 0.36695230 0.5457190