Compares paired-samples standardized mean differences in original and follow-up studies

This function computes confidence intervals for a paired-samples standardized mean difference from an original study and a follow-up study. Confidence intervals for the difference and average effect size also are computed. The same results can be obtained using the meta.lc.stdmean.ps function with appropriate contrast coefficients. The confidence level for the difference is 1 – 2alpha.

replicate.stdmean.ps(
  alpha,
  m11,
  m12,
  sd11,
  sd12,
  cor1,
  n1,
  m21,
  m22,
  sd21,
  sd22,
  cor2,
  n2
)

Arguments

alpha
alpha level for 1-alpha confidence

m11
sample mean for group 1 in original study

m12
sample mean for group 2 in original study

sd11
sample SD for group 1 in original study

sd12
sample SD for group 2 in original study

cor1
sample correlation of paired observations in orginal study

n1
sample size in original study

m21
sample mean for group 1 in follow-up study

m22
sample mean for group 2 in follow-up study

sd21
sample SD for group 1 in follow-up study

sd22
sample SD for group 2 in follow-up study

cor2
sample correlation of paired observations in follow-up study

n2
sample size in follow-up study

Value

A 4-row matrix. The rows are:

Row 1 summarizes the Original study

Row 2 summarizes the Follow-up study

Row 3 estimates the difference between studies

Row 4 estimates the average effect size between the two studies

The columns are:

Estimate - effect size estimate

SE - standard error

LL - lower limit of the confidence interval

UL - upper limit of the confidence interval

References

Bonett DG (2021). “Design and analysis of replication studies.” Organizational Research Methods, 24(3), 513--529. ISSN 1094-4281, doi: 10.1177/1094428120911088 , https://journals.sagepub.com/doi/10.1177/1094428120911088.

Examples

replicate.stdmean.ps(
  alpha = .05, 
  m11 = 86.22, m12 = 70.93, sd11 = 14.89, sd12 = 12.32, cor1 = .765, n1 = 20, 
  m21 = 84.81, m22 = 77.24, sd21 = 15.68, sd22 = 16.95, cor2 = .702, n2 = 75
)

#>                        Estimate         SE        LL        UL
#> Orginal:              1.0890300 0.22915553 0.6697353 1.5680085
#> Follow-up:            0.4604958 0.09590506 0.2756687 0.6516096
#> Original - Follow-up: 0.6552328 0.24841505 0.2466264 1.0638392
#> Average:              0.7747629 0.12420752 0.5313206 1.0182052

# Should return:
#                         Estimate         SE        LL        UL
#  Orginal:              1.0890300 0.22915553 0.6697353 1.5680085
#  Follow-up:            0.4604958 0.09590506 0.2756687 0.6516096
#  Original - Follow-up: 0.6552328 0.24841505 0.2466264 1.0638392
#  Average:              0.7747629 0.12420752 0.5313206 1.0182052