Compares 2-group standardized mean differences in original and follow-up studies

This function computes confidence intervals for a 2-group standardized mean difference from an original study and a follow-up study. Confidence intervals for the difference and average effect size also are computed. The same results can be obtained using the meta.lc.stdmean2 function with appopriate contrast coefficients. The confidence level for the difference is 1 – 2alpha.

replicate.stdmean2(
  alpha,
  m11,
  m12,
  sd11,
  sd12,
  n11,
  n12,
  m21,
  m22,
  sd21,
  sd22,
  n21,
  n22
)

Arguments

alpha

alpha level for 1-alpha confidence

m11

estimated mean for group 1 in original study

m12

estimated mean for group 2 in original study

sd11

estimated SD for group 1 in original study

sd12

estimated SD for group 2 in original study

n11

sample size for group 1 in original study

n12

sample size for group 2 in original study

m21

estimated mean for group 1 in follow-up study

m22

estimated mean for group 2 in follow-up study

sd21

estimated SD for group 1 in follow-up study

sd22

estimated SD for group 2 in follow-up study

n21

sample size for group 1 in follow-up study

n22

sample size for group 2 in follow-up study

Value

A 4-row matrix. The rows are:

Row 1 summarizes the original study

Row 2 summarizes the follow-up study

Row 3 estimates the difference between studies

Row 4 estimates the average effect size between the two studies

The columns are:

Estimate - effect size estimate

SE - standard error

LL - lower limit of the confidence interval

UL - upper limit of the confidence interval

References

Bonett DG (2021). “Design and analysis of replication studies.” Organizational Research Methods, 24(3), 513--529. ISSN 1094-4281, doi: 10.1177/1094428120911088 , https://journals.sagepub.com/doi/10.1177/1094428120911088.

Examples

replicate.stdmean2(.05, 21.9, 16.1, 3.82, 3.21, 40, 40, 
                        25.2, 19.1, 3.98, 3.79, 75, 75)
#>                         Estimate        SE         LL        UL
#> Original:             1.62803662 0.2629049  1.1286100 2.1591783
#> Follow-up:            1.56170447 0.1880713  1.2010594 1.9382853
#> Original - Follow-up: 0.07422178 0.3232488 -0.4574752 0.6059188
#> Average:              1.59487055 0.1616244  1.2780925 1.9116486

# Should return: 
#                           Estimate        SE         LL        UL
#  Original:              1.62803662 0.2594668  1.1353486 2.1524396
#  Follow-up:             1.56170447 0.1870576  1.2030461 1.9362986
#  Original - Follow-up:  0.07422178 0.3198649 -0.4519092 0.6003527
#  Average:               1.59487055 0.1599325  1.2814087 1.9083324