Computes the standard error for a Pearson or partial correlation

This function can be used to compute the standard error of a Pearson or partial correlation using the estimated correlation, sample size, and number of control variables. The correlation, along with the standard error output from this function, can be used as input in the meta.ave.gen function in applications where a combination of different types of correlations are used in the meta-analysis.

se.cor(cor, s, n)

Arguments

cor

estimated Pearson or partial correlation

s

number of control variables (set to 0 for Pearson)

n

sample size

Value

Returns a one-row matrix:

Estimate - Pearson or partial correlation (from input)

SE - standard error

References

Bonett DG (2008). “Meta-analytic interval estimation for bivariate correlations.” Psychological Methods, 13(3), 173--181. ISSN 1939-1463, doi:10.1037/a0012868 .

Examples

se.cor(.40, 0, 55)
#>               Estimate       SE
#> Correlation:       0.4 0.116487

# Should return: 
#               Estimate       SE
# Correlation:       0.4 0.116487